+++ app/src/main/java/com/semicolon/xquare_android/MainActivity.kt
@@ -1,5 +1,8 @@
-package com.semicolon.xquare_android.ui
+package com.semicolon.xquare_android
 
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.material.rememberScaffoldState
@@ -9,7 +12,30 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.semicolon.xquare_android.ui.navigation.BottomNavigationItem
+import com.semicolon.xquare_android.navigation.AppNavigationItem
+import com.semicolon.xquare_android.navigation.BottomNavigationItem
+import com.semicolon.xquare_android.theme.XquareandroidTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            XquareandroidTheme {
+                BaseApp()
+            }
+        }
+    }
+}
+
+@Composable
+fun BaseApp() {
+    val navController = rememberNavController()
+    NavHost(navController = navController, startDestination = AppNavigationItem.Main.route) {
+        composable(AppNavigationItem.Main.route) {
+            Main(navController)
+        }
+    }
+}
 
 @Composable
 fun Main(navController: NavController) {
