+++ app/src/main/java/com/semicolon/xquare_android/navigation/AppNavigationItem.kt
@@ -1,4 +1,4 @@
-package com.semicolon.xquare_android.ui.navigation
+package com.semicolon.xquare_android.navigation
 
 enum class AppNavigationItem(val route: String) {
     Main("main")
