+++ app/src/main/java/com/semicolon/xquare_android/navigation/BottomNavigationItem.kt
@@ -1,4 +1,4 @@
-package com.semicolon.xquare_android.ui.navigation
+package com.semicolon.xquare_android.navigation
 
 enum class BottomNavigationItem(val route: String) {
     Home("home")
