+++ app/src/main/java/com/semicolon/xquare_android/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.semicolon.xquare_android.ui.theme
+package com.semicolon.xquare_android.theme
 
 import androidx.compose.ui.graphics.Color
